requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
passlib
python-jose
python-multipart
